+++ lib/Secucard.Connect/Product/Service/Services.cs
@@ -5,5 +5,7 @@ namespace Secucard.Connect.Product.Service
         public IdentRequestsService Identrequests { get; set; }
 
         public IdentResultsService Identresults { get; set; }
+
+        public UploadidentsService Uploadidents { get; set; }
     }
 }